implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in CV

+++ Models/CheckVoucherHeader.cs
@@ -9,7 +9,7 @@ namespace Accounting_System.Models
         [Display(Name = "CV No")]
         public string? CVNo { get; set; }
 
-        public int SeriesNumber { get; set; }
+        public long SeriesNumber { get; set; }
 
         public DateTime Date { get; set; }
